un-common server and cli config
